+++ ydb/core/config/validation/ya.make
@@ -2,6 +2,7 @@ LIBRARY()
 
 SRCS(
     validators.h
+    validators.cpp
 )
 
 PEERDIR(
